Added earlt verification to deposit_event for TokenMessageData
